[grsec]: go unofficial...
As announced in [1] no more grsecurity or PaX patches will be publicly
available. This is an attempt to keep, at least, the v4.9 port alive for
the time being.

To comply with the trademark compliance request of [2], the grsecurity
logo has been removed and the local version -- the uname suffix --
altered to '-unofficial_grsec' which should make it clearly visible to
users that this is *not* the official grsecurity patch set but an
unofficial port.

May it help the community which depends on the security features of the
original patch for the life time of the v4.9 Linux kernel.

[1] https://grsecurity.net/passing_the_baton.php
[2] https://grsecurity.net/passing_the_baton_faq.php

Signed-off-by: Mathias Krause <minipli@googlemail.com>