firmware: drop bnx2 firmware files added by grsec
Even though the bnx2.ko network driver depends on those files for some
of its supported NICs, they don't provide any additional security for
the kernel itself, so fall out of scope for the overall goal of the
grsec(tm) patch. It's rather just a usability issue, not being dependent
on an external firmware package. However, vanilla Linux has the same
issue and distros can handle this just fine. So drop those hunks.

If you're depending on this driver and, in turn, on those firmware
files, look for a firmware package called firmware-bnx2, linux-firmware
or similar or get the required files directly from the linux-firmware
git repo at [1].

Reported-by: HacKurx in https://github.com/minipli/linux-unofficial_grsec/issues/7
Signed-off-by: Mathias Krause <minipli@googlemail.com>

[1] https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git